+++ nextjs/src/app/reports/[id]/(components)/MapLayers/UKWards.tsx
@@ -2,11 +2,14 @@ import { AnalyticalAreaType, GroupedDataCount } from '@/__generated__/graphql'
 import { useLoadedMap } from '@/lib/map'
 import { useEffect, useState } from 'react'
 import { Layer, Source } from 'react-map-gl'
+import { addCountByGssToMapboxLayer } from '../../addCountByGssToMapboxLayer'
+import {
+  getChoroplethColours,
+  getChoroplethEdge,
+} from '../../getChoroplethStyles'
+import { Tileset } from '../../types'
+import useBoundaryAnalytics from '../../useBoundaryAnalytics'
 import { useReport } from '../ReportProvider'
-import { addCountByGssToMapboxLayer } from './addCountByGssToMapboxLayer'
-import { getChoroplethColours, getChoroplethEdge } from './getChoroplethStyles'
-import { Tileset } from './types'
-import useDataSources from './useDataSources'
 
 // https://studio.mapbox.com/tilesets/commonknowledge.0rzbo365
 function getTileset(data: GroupedDataCount[]): Tileset {
@@ -23,7 +26,10 @@ function getTileset(data: GroupedDataCount[]): Tileset {
 
 const UKWards = () => {
   const { report } = useReport()
-  const countsByWard = useDataSources(report, AnalyticalAreaType.AdminWard)
+  const countsByWard = useBoundaryAnalytics(
+    report,
+    AnalyticalAreaType.AdminWard
+  )
   const map = useLoadedMap()
   const [tileset, setTileset] = useState<Tileset | null>(null)
   const visibility =
