+++ nextjs/src/app/reports/[id]/addCountByGssToMapboxLayer.ts
