+++ nextjs/src/app/reports/[id]/getChoroplethStyles.ts
