+++ nextjs/src/app/reports/[id]/types.ts
